[WIP] Replace ::slate with ::node spec
Refactor ::slate to leverage spec more, specifically encoding the structure of a
node in a spec itself rather than a function that returns a regex spec.
Currently a WIP because I did not replace where ::slate is used (`make-ast`) as
the ::node spec returns a different conform value than ::slate --- the ::node
spec is more granular with its labeling (i.e. gives a more verbose path) but
it's structure is not a drop-in replacement for ::slate.

Eventually ::node should replace the node function and the ::slate spec.